Added more notes and orgamized the files again
