--- SRMP/PauseState.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SRMultiplayer
-{
-    public enum PauseState
-    {
-        Pause,
-        Playing
-    }
-}
